carla_stream-main/backend/pyproject.toml
[tool.poetry]
name = "wgstream"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.10.0.84"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
mss = "^9.0.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
